Se agrega la clase validación que permite controlar que datos ingresados esten completos

+++ InterlocutorComercial/InterlocutorComercial.cpp
@@ -27,7 +27,7 @@ InterlocutorComercial::InterlocutorComercial(string nombreArchivo)
 }
 InterlocutorComercial::~InterlocutorComercial()
 {}
-void InterlocutorComercial::setId_ic(long id_ic)
+void InterlocutorComercial::setId_ic(int id_ic)
 {
 	this->id_ic = id_ic;
 }
@@ -39,7 +39,7 @@ void InterlocutorComercial::setApellido(string apellido)
 {
 	strcpy(this->apellido, apellido.c_str());
 }
-void InterlocutorComercial::setDni(long dni)
+void InterlocutorComercial::setDni(int dni)
 {
 	this->dni = dni;
 }
@@ -55,7 +55,7 @@ void InterlocutorComercial::setDireccionId(Direccion& direccionId)
 {
 	this->direccionID = direccionId;
 }
-void InterlocutorComercial::setNumPosicionArchivo(long numPosArch)
+void InterlocutorComercial::setNumPosicionArchivo(int numPosArch)
 {
 	this->numPosicionArch = numPosArch;
 }
@@ -67,7 +67,7 @@ void InterlocutorComercial::setActivo(bool activo)
 {
 	this->activo = activo;
 }
-long InterlocutorComercial::getId_ic()
+int InterlocutorComercial::getId_ic()
 {
 	return this->id_ic;
 }
@@ -79,7 +79,7 @@ string InterlocutorComercial::getApellido()
 {
 	return this->apellido;
 }
-long InterlocutorComercial::getDni()
+int InterlocutorComercial::getDni()
 {
 	return this->dni;
 }
@@ -95,7 +95,7 @@ Direccion& InterlocutorComercial::getDireccionId()
 {
 	return this->direccionID;
 }
-long InterlocutorComercial::getNumPosicionArhivo()
+int InterlocutorComercial::getNumPosicionArhivo()
 {
 	return this->numPosicionArch;
 }
